requirements.txt
# Core dependencies
pyswisseph>=2.10.0
numpy>=1.21.0
pydantic>=2.0.0
pytz>=2023.3

# API server dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
email-validator>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
redis>=4.6.0
prometheus-client>=0.17.0
httpx>=0.24.0
websockets>=11.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=23.1.0

# Additional production dependencies
requests>=2.31.0
PyJWT>=2.8.0 